+++ restaurant-service/src/main/java/com/example/restaurantservice/model/Transaction.java
@@ -17,12 +17,8 @@ public class Transaction {
     private Long id;
 
     private Long restaurantId;
-
     private LocalDate date;
-
     private String description;
-
     private String bankName;
-
-    private Double amount;
+    private Long amount;
 }
